export.sh: Store log messages in export directory and refactor operations to avoid warnings.

+++ export.sh
@@ -3,21 +3,29 @@
 # export current working directory in a format suitable for sending to
 # MySQL as a snapshot.
 
+set -eu
+
+if [ $# -ne 1 ] ; then
+  echo "Usage: export.sh revision-number-of-last-snapshot"
+  exit 1
+fi
+
 rm -rf to-mysql
-svn export . to-mysql
+mkdir -p to-mysql/storage/
+svn log -v -r "$1:BASE" > to-mysql/log
+svn export . to-mysql/storage/innobase
 cd to-mysql
 
-mkdir innobase
-mv * innobase
 mkdir -p sql mysql-test/t mysql-test/r mysql-test/include
-cd innobase
+cd storage/innobase
 
-mv handler/* ../sql
+mv handler/* ../../sql
 rmdir handler
 
-mv mysql-test/*.test mysql-test/*.opt ../mysql-test/t
-mv mysql-test/*.result ../mysql-test/r
-mv mysql-test/*.inc ../mysql-test/include
+mv mysql-test/*.test mysql-test/*.opt ../../mysql-test/t
+mv mysql-test/*.result ../../mysql-test/r
+mv mysql-test/*.inc ../../mysql-test/include
 rmdir mysql-test
 
-rm setup.sh export.sh
+rm setup.sh export.sh compile-innodb-debug
+
